98: rename text.js to fields.js
