+++ src/components/Card/AdsorbateCard/DischargeLimit.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import {BodyText, FieldLabel, FieldValue} from "../CommonCardStyles";
 import {spillLimit} from "../../../common/formatting/spillLimit";
-import {ADSORBATE_FIELDS} from "../../../common/text";
+import {ADSORBATE_FIELDS} from "../../../common/fields";
 
 export const DischargeLimit = ({dischargeLimit}) => {
   return (
