+++ src/components/Card/AdsorbentCard/Ph.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import {BodyText, FieldLabel, FieldValue} from "../CommonCardStyles";
-import {ADSORBENT_FIELDS} from "../../../common/text";
+import {ADSORBENT_FIELDS} from "../../../common/fields";
 
 export const Ph = ({ph}) => {
   return (
