+++ src/components/Card/AdsorbentCard/SBetValue.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import {BodyText, FieldLabel, FieldValue} from "../CommonCardStyles";
 import {Sup} from "../../Card/CommonCardStyles";
-import {ADSORBENT_FIELDS} from "../../../common/text";
+import {ADSORBENT_FIELDS} from "../../../common/fields";
 
 export const SBetValue = ({sBet}) => {
   return (
