+++ src/components/Card/AdsorbentCard/VBetValue.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import {BodyText, FieldLabel, FieldValue} from "../CommonCardStyles";
 import {Sup} from "../CommonCardStyles";
-import {ADSORBENT_FIELDS} from "../../../common/text";
+import {ADSORBENT_FIELDS} from "../../../common/fields";
 
 export const VBetValue = ({vBet}) => {
   return (
