+++ src/components/Search/Fields/IonicCharge.js
@@ -3,7 +3,7 @@ import TextField from "@material-ui/core/TextField";
 import React from "react";
 import {getTextFieldStyles} from "../Styles";
 import {useScreenWidth} from "../../../customHooks/useScreenWidth";
-import {ADSORBATE_FIELDS} from "../../../common/text";
+import {ADSORBATE_FIELDS} from "../../../common/fields";
 
 export const IonicCharge = ({charge, handleChargeChange, onKeyPress}) => {
   const responsiveWidth = 848;
